# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MCG Building Violation Management System API
  description: REST API consumed by the MCG web portal module and the MCG HARYANA mobile app (building-violations module).
  version: 2.0.0
paths:
...
  /building-violations/api/reports/:
    get:
      tags:
      - building-violations
      summary: Report List
      operationId: report_list_building_violations_api_reports__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
  /building-violations/api/reports/{name}/:
    get:
      tags:
      - building-violations
      summary: Run Report
      operationId: run_report_building_violations_api_reports__name___get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
          title: Name
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /building-violations/api/auth/otp/verify/
servers:
- url: /
  description: Same origin (the module is mounted at /building-violations/)